Added Tilesets
The two tilesets that we'll need to build the level are now finished.
Use the Floor tileset to create the areas that the player can walk on,
and the Wall tileset to create areas they can't.

In addition, I've created separate event sheets for each of us (the area
where all the code goes), so that none of our code changes will clash
when pushed.