Last Commit of 1/30
Almost done, the rest will be easy to implement tomorrow. Map is mostly
designed, with the only required alterations being the addition of the
altar and putting down some more carpet.

Adding the batteries and a random spawner for them should be easy, as
should the gun and the zombie spawner (which works based upon the same
principle.)

I think we're well on track, though tomorrow may be a slight crunch.
I'll be getting up at 7 to work at RPI, and we leave for the Jam at 11
AM.