stuff
Made a battery image and added some possible events to my sheet for
turning the flashlight off when it's out of charge and converting
batteries to charge on collide with them. Maybe active interaction to
pick up would be better? Not sure if there should be max charge for the
light.